cleaned up broker_feature
