+++ package.json
@@ -28,6 +28,9 @@
   "dependencies": {
     "cheerio": "^1.1.2",
     "node-fetch": "^3.3.2",
-    "puppeteer": "^24.14.0"
+    "puppeteer": "^24.14.0",
+    "puppeteer-extra": "^3.3.6",
+    "puppeteer-extra-plugin-stealth": "^2.11.2",
+    "puppeteer-real-browser": "^1.4.3"
   }
 }
